Trying out some audio filtering...

+++ winnes/waveout.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 // Returns NULL if all buffers are full
-int8_t* waveout_get_current_buffer();
+int16_t* waveout_get_current_buffer();
 void waveout_queue_buffer();
 
 void waveout_initialize(unsigned int sample_rate, unsigned buffer_size);
